requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Nyaya-Setu — GPU Requirements (NVIDIA RTX 4050 / CUDA 12.x)
# ─────────────────────────────────────────────────────────────────────────────
#
# INSTALL ORDER — run these EXACTLY in this sequence:
#
#   Step 1 (CUDA PyTorch — must be first):
#     pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
#   Step 2 (everything else):
#     pip install -r requirements.txt
#
#   Step 3 (verify GPU is detected):
#     python -c "import torch; print(torch.cuda.get_device_name(0))"
#     Expected: NVIDIA GeForce RTX 4050 Laptop GPU
#
# ─────────────────────────────────────────────────────────────────────────────

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.7
chromadb==0.5.23
sentence-transformers==3.3.1
rank-bm25==0.2.2
ollama==0.4.2
PyMuPDF==1.24.14
Pillow==11.0.0
reportlab==4.2.5
requests==2.32.3
pydub==0.25.1
librosa==0.10.2
scipy==1.14.1
soundfile==0.12.1
transformers==4.47.1
accelerate==1.2.1
datasets==3.2.0
twilio==9.4.0
python-dotenv==1.0.1
pydantic==2.10.3
tqdm==4.67.1
numpy==1.26.4
googletrans==4.0.0-rc1
pytesseract==0.3.13